modularising code and seperating date and mrn filters into components so that they can be rendered conditionally
